a version for third times of large-scale evaluation

+++ strategy/tree.h
@@ -17,7 +17,7 @@ struct SpaceTreeNode
     int dim_num = 0;
 
     // Number of active addresses in this dimension
-    float active = 0;
+    uint64_t active = 0;
 };
 
 typedef std::vector<struct SpaceTreeNode*> Node_List;
